requirements.txt
tensorflow>=2
tensorflow-io
tensorflow-text
tqdm
pyyaml
pydantic
